+++ src/components/NewMeeting.js
@@ -1,5 +1,5 @@
 import "../css/newMeeting.css";
-import EventCreation from './EventCreation'
+import EventCreation from './EventCreation/EventCreateion'
 
 
 export default function NewMeeting() {
